REFACTOR: move Operator interface
