+++ core/src/main/kotlin/uk/dioxic/mgenerate/codecs/OperatorExecutionCodec.kt
@@ -6,7 +6,7 @@ import org.bson.codecs.Codec
 import org.bson.codecs.DecoderContext
 import org.bson.codecs.EncoderContext
 import org.bson.codecs.configuration.CodecRegistry
-import uk.dioxic.mgenerate.annotations.Operator
+import uk.dioxic.mgenerate.operators.Operator
 
 class OperatorExecutionCodec(
     private val registry: CodecRegistry
