+++ core/src/main/kotlin/uk/dioxic/mgenerate/codecs/OperatorExecutionCodecProvider.kt
@@ -3,7 +3,7 @@ package uk.dioxic.mgenerate.codecs
 import org.bson.codecs.Codec
 import org.bson.codecs.configuration.CodecProvider
 import org.bson.codecs.configuration.CodecRegistry
-import uk.dioxic.mgenerate.annotations.Operator
+import uk.dioxic.mgenerate.operators.Operator
 
 class OperatorExecutionCodecProvider: CodecProvider {
