+++ core/src/main/kotlin/uk/dioxic/mgenerate/operators/StringOperators.kt
@@ -1,7 +1,6 @@
 package uk.dioxic.mgenerate.operators
 
 import uk.dioxic.mgenerate.annotations.Alias
-import uk.dioxic.mgenerate.annotations.Operator
 
 @Alias("concat", "join")
 class Concat(
