+++ core/src/main/kotlin/uk/dioxic/mgenerate/operators/general/ChooseOperator.kt
@@ -1,8 +1,8 @@
 package uk.dioxic.mgenerate.operators.general
 
 import uk.dioxic.mgenerate.annotations.Alias
-import uk.dioxic.mgenerate.annotations.Operator
 import uk.dioxic.mgenerate.extensions.nextElement
+import uk.dioxic.mgenerate.operators.Operator
 import kotlin.random.Random
 
 @Alias("choose")
