+++ core/src/main/kotlin/uk/dioxic/mgenerate/operators/general/ObjectIdOperator.kt
@@ -2,7 +2,7 @@ package uk.dioxic.mgenerate.operators.general
 
 import org.bson.types.ObjectId
 import uk.dioxic.mgenerate.annotations.Alias
-import uk.dioxic.mgenerate.annotations.Operator
+import uk.dioxic.mgenerate.operators.Operator
 
 @Alias("objectId")
 class ObjectIdOperator : Operator<ObjectId> {
