+++ core/src/main/kotlin/uk/dioxic/mgenerate/operators/general/PickSetOperator.kt
@@ -2,8 +2,8 @@ package uk.dioxic.mgenerate.operators.general
 
 import org.apache.logging.log4j.kotlin.logger
 import uk.dioxic.mgenerate.annotations.Alias
-import uk.dioxic.mgenerate.annotations.Operator
 import uk.dioxic.mgenerate.extensions.nextElement
+import uk.dioxic.mgenerate.operators.Operator
 import kotlin.math.min
 import kotlin.random.Random
